refactor(components): [table] make component generic
