+++ packages/components/table/index.ts
@@ -2,6 +2,8 @@ import { withInstall, withNoopInstall } from '@element-plus/utils'
 import Table from './src/table.vue'
 import TableColumn from './src/tableColumn'
 
+import type { ComponentInstance } from 'vue'
+import type { ComponentExposed } from 'vue-component-type-helpers'
 import type { SFCWithInstall } from '@element-plus/utils'
 
 export const ElTable: SFCWithInstall<typeof Table> & {
@@ -13,9 +15,11 @@ export default ElTable
 export const ElTableColumn: SFCWithInstall<typeof TableColumn> =
   withNoopInstall(TableColumn)
 
-export type TableInstance = InstanceType<typeof Table> & unknown
+export type TableInstance = ComponentInstance<typeof Table> &
+  ComponentExposed<typeof Table>
 
-export type TableColumnInstance = InstanceType<typeof TableColumn> & unknown
+export type TableColumnInstance = ComponentInstance<typeof TableColumn> &
+  ComponentExposed<typeof TableColumn>
 
 export type {
   SummaryMethod,
