+++ packages/components/table/src/table.vue
@@ -170,16 +170,10 @@
   </div>
 </template>
 
-<script lang="ts">
-import {
-  computed,
-  defineComponent,
-  getCurrentInstance,
-  onBeforeUnmount,
-  provide,
-} from 'vue'
+<script lang="ts" setup generic="T extends DefaultRow = DefaultRow">
+import { computed, getCurrentInstance, onBeforeUnmount, provide } from 'vue'
 import { debounce } from 'lodash-unified'
-import { Mousewheel } from '@element-plus/directives'
+import { Mousewheel as vMousewheel } from '@element-plus/directives'
 import { useLocale, useNamespace } from '@element-plus/hooks'
 import { useGlobalConfig } from '@element-plus/components/config-provider'
 import ElScrollbar from '@element-plus/components/scrollbar'
@@ -188,235 +182,239 @@ import TableLayout from './table-layout'
 import TableHeader from './table-header'
 import TableBody from './table-body'
 import TableFooter from './table-footer'
+import { createTableId } from './util'
 import useUtils from './table/utils-helper'
 import { convertToRows } from './table-header/utils-helper'
 import useStyle from './table/style-helper'
 import useKeyRender from './table/key-render-helper'
-import defaultProps from './table/defaults'
 import { TABLE_INJECTION_KEY } from './tokens'
 import { hColgroup } from './h-helper'
 import { useScrollbar } from './composables/use-scrollbar'
 
-import type { Table } from './table/defaults'
+import type { DefaultRow, Table, TableProps } from './table/defaults'
 
-let tableIdSeed = 1
-export default defineComponent({
+defineOptions({
   name: 'ElTable',
-  directives: {
-    Mousewheel,
-  },
-  components: {
-    TableHeader,
-    TableBody,
-    TableFooter,
-    ElScrollbar,
-    hColgroup,
-  },
-  props: defaultProps,
-  emits: [
-    'select',
-    'select-all',
-    'selection-change',
-    'cell-mouse-enter',
-    'cell-mouse-leave',
-    'cell-contextmenu',
-    'cell-click',
-    'cell-dblclick',
-    'row-click',
-    'row-contextmenu',
-    'row-dblclick',
-    'header-click',
-    'header-contextmenu',
-    'sort-change',
-    'filter-change',
-    'current-change',
-    'header-dragend',
-    'expand-change',
-    'scroll',
-  ],
-  setup(props) {
-    type Row = (typeof props.data)[number]
-    const { t } = useLocale()
-    const ns = useNamespace('table')
-    const globalConfig = useGlobalConfig('table')
-    const table = getCurrentInstance() as Table<Row>
-    provide(TABLE_INJECTION_KEY, table)
-    const store = createStore<Row>(table, props)
-    table.store = store
-    const layout = new TableLayout<Row>({
-      store: table.store,
-      table,
-      fit: props.fit,
-      showHeader: props.showHeader,
-    })
-    table.layout = layout
+})
 
-    const isEmpty = computed(() => (store.states.data.value || []).length === 0)
+const props = withDefaults(defineProps<TableProps<T>>(), {
+  data: () => [],
+  fit: true,
+  showHeader: true,
+  selectOnIndeterminate: true,
+  indent: 16,
+  treeProps: () => ({
+    hasChildren: 'hasChildren',
+    children: 'children',
+    checkStrictly: false,
+  }),
+  style: () => ({}),
+  className: '',
+  tableLayout: 'fixed',
+  showOverflowTooltip: undefined,
+  scrollbarTabindex: undefined,
+  allowDragLastColumn: true,
+})
 
-    /**
-     * open functions
-     */
-    const {
-      setCurrentRow,
-      getSelectionRows,
-      toggleRowSelection,
-      clearSelection,
-      clearFilter,
-      toggleAllSelection,
-      toggleRowExpansion,
-      clearSort,
-      sort,
-      updateKeyChildren,
-    } = useUtils<Row>(store)
-    const {
-      isHidden,
-      renderExpanded,
-      setDragVisible,
-      isGroup,
-      handleMouseLeave,
-      handleHeaderFooterMousewheel,
-      tableSize,
-      emptyBlockStyle,
-      resizeProxyVisible,
-      bodyWidth,
-      resizeState,
-      doLayout,
-      tableBodyStyles,
-      tableLayout,
-      scrollbarViewStyle,
-      scrollbarStyle,
-    } = useStyle<Row>(props, layout, store, table)
+defineEmits([
+  'select',
+  'select-all',
+  'selection-change',
+  'cell-mouse-enter',
+  'cell-mouse-leave',
+  'cell-contextmenu',
+  'cell-click',
+  'cell-dblclick',
+  'row-click',
+  'row-contextmenu',
+  'row-dblclick',
+  'header-click',
+  'header-contextmenu',
+  'sort-change',
+  'filter-change',
+  'current-change',
+  'header-dragend',
+  'expand-change',
+  'scroll',
+])
 
-    const { scrollBarRef, scrollTo, setScrollLeft, setScrollTop } =
-      useScrollbar()
+const { t } = useLocale()
+const ns = useNamespace('table')
+const globalConfig = useGlobalConfig('table')
+const table = getCurrentInstance() as Table<T>
+provide(TABLE_INJECTION_KEY, table)
+const store = createStore<T>(table, props)
+table.store = store
+const layout = new TableLayout<T>({
+  store: table.store,
+  table,
+  fit: props.fit,
+  showHeader: props.showHeader,
+})
+table.layout = layout
 
-    const debouncedUpdateLayout = debounce(doLayout, 50)
+const isEmpty = computed(() => (store.states.data.value || []).length === 0)
 
-    const tableId = `${ns.namespace.value}-table_${tableIdSeed++}`
-    table.tableId = tableId
-    table.state = {
-      isGroup,
-      resizeState,
-      doLayout,
-      debouncedUpdateLayout,
-    }
-    const computedSumText = computed(
-      () => props.sumText ?? t('el.table.sumText')
-    )
+/**
+ * open functions
+ */
+const {
+  setCurrentRow,
+  getSelectionRows,
+  toggleRowSelection,
+  clearSelection,
+  clearFilter,
+  toggleAllSelection,
+  toggleRowExpansion,
+  clearSort,
+  sort,
+  updateKeyChildren,
+} = useUtils<T>(store)
+const {
+  isHidden,
+  renderExpanded,
+  setDragVisible,
+  isGroup,
+  handleMouseLeave,
+  handleHeaderFooterMousewheel,
+  tableSize,
+  emptyBlockStyle,
+  resizeProxyVisible,
+  bodyWidth,
+  resizeState,
+  doLayout,
+  tableBodyStyles,
+  tableLayout,
+  scrollbarViewStyle,
+  scrollbarStyle,
+} = useStyle<T>(props, layout, store, table)
 
-    const computedEmptyText = computed(() => {
-      return props.emptyText ?? t('el.table.emptyText')
-    })
+const { scrollBarRef, scrollTo, setScrollLeft, setScrollTop } = useScrollbar()
 
-    const computedTooltipEffect = computed(
-      () => props.tooltipEffect ?? globalConfig.value?.tooltipEffect
-    )
+const debouncedUpdateLayout = debounce(doLayout, 50)
 
-    const computedTooltipOptions = computed(
-      () => props.tooltipOptions ?? globalConfig.value?.tooltipOptions
-    )
+const tableId = createTableId(ns.namespace.value)
+const context = table
+table.tableId = tableId
+table.state = {
+  isGroup,
+  resizeState,
+  doLayout,
+  debouncedUpdateLayout,
+}
+const computedSumText = computed(() => props.sumText ?? t('el.table.sumText'))
 
-    const columns = computed(() => {
-      return convertToRows(store.states.originColumns.value)[0]
-    })
+const computedEmptyText = computed(() => {
+  return props.emptyText ?? t('el.table.emptyText')
+})
 
-    useKeyRender(table)
+const computedTooltipEffect = computed(
+  () => props.tooltipEffect ?? globalConfig.value?.tooltipEffect
+)
 
-    onBeforeUnmount(() => {
-      debouncedUpdateLayout.cancel()
-    })
+const computedTooltipOptions = computed(
+  () => props.tooltipOptions ?? globalConfig.value?.tooltipOptions
+)
 
-    return {
-      ns,
-      layout,
-      store,
-      columns,
-      handleHeaderFooterMousewheel,
-      handleMouseLeave,
-      tableId,
-      tableSize,
-      isHidden,
-      isEmpty,
-      renderExpanded,
-      resizeProxyVisible,
-      resizeState,
-      isGroup,
-      bodyWidth,
-      tableBodyStyles,
-      emptyBlockStyle,
-      debouncedUpdateLayout,
-      /**
-       * @description used in single selection Table, set a certain row selected. If called without any parameter, it will clear selection
-       */
-      setCurrentRow,
-      /**
-       * @description returns the currently selected rows
-       */
-      getSelectionRows,
-      /**
-       * @description used in multiple selection Table, toggle if a certain row is selected. With the second parameter, you can directly set if this row is selected
-       */
-      toggleRowSelection,
-      /**
-       * @description used in multiple selection Table, clear user selection
-       */
-      clearSelection,
-      /**
-       * @description clear filters of the columns whose `columnKey` are passed in. If no params, clear all filters
-       */
-      clearFilter,
-      /**
-       * @description used in multiple selection Table, toggle select all and deselect all
-       */
-      toggleAllSelection,
-      /**
-       * @description used in expandable Table or tree Table, toggle if a certain row is expanded. With the second parameter, you can directly set if this row is expanded or collapsed
-       */
-      toggleRowExpansion,
-      /**
-       * @description clear sorting, restore data to the original order
-       */
-      clearSort,
-      /**
-       * @description refresh the layout of Table. When the visibility of Table changes, you may need to call this method to get a correct layout
-       */
-      doLayout,
-      /**
-       * @description sort Table manually. Property `prop` is used to set sort column, property `order` is used to set sort order
-       */
-      sort,
-      /**
-       * @description used in lazy Table, must set `rowKey`, update key children
-       */
-      updateKeyChildren,
-      t,
-      setDragVisible,
-      context: table,
-      computedSumText,
-      computedEmptyText,
-      computedTooltipEffect,
-      computedTooltipOptions,
-      tableLayout,
-      scrollbarViewStyle,
-      scrollbarStyle,
-      scrollBarRef,
-      /**
-       * @description scrolls to a particular set of coordinates
-       */
-      scrollTo,
-      /**
-       * @description set horizontal scroll position
-       */
-      setScrollLeft,
-      /**
-       * @description set vertical scroll position
-       */
-      setScrollTop,
-      /**
-       * @description whether to allow drag the last column
-       */
-      allowDragLastColumn: props.allowDragLastColumn,
-    }
-  },
+const columns = computed(() => {
+  return convertToRows(store.states.originColumns.value)[0]
+})
+
+useKeyRender(table)
+
+onBeforeUnmount(() => {
+  debouncedUpdateLayout.cancel()
+})
+
+defineExpose({
+  ns,
+  layout,
+  store,
+  columns,
+  handleHeaderFooterMousewheel,
+  handleMouseLeave,
+  tableId,
+  tableSize,
+  isHidden,
+  isEmpty,
+  renderExpanded,
+  resizeProxyVisible,
+  resizeState,
+  isGroup,
+  bodyWidth,
+  tableBodyStyles,
+  emptyBlockStyle,
+  debouncedUpdateLayout,
+  /**
+   * @description used in single selection Table, set a certain row selected. If called without any parameter, it will clear selection
+   */
+  setCurrentRow,
+  /**
+   * @description returns the currently selected rows
+   */
+  getSelectionRows,
+  /**
+   * @description used in multiple selection Table, toggle if a certain row is selected. With the second parameter, you can directly set if this row is selected
+   */
+  toggleRowSelection,
+  /**
+   * @description used in multiple selection Table, clear user selection
+   */
+  clearSelection,
+  /**
+   * @description clear filters of the columns whose `columnKey` are passed in. If no params, clear all filters
+   */
+  clearFilter,
+  /**
+   * @description used in multiple selection Table, toggle select all and deselect all
+   */
+  toggleAllSelection,
+  /**
+   * @description used in expandable Table or tree Table, toggle if a certain row is expanded. With the second parameter, you can directly set if this row is expanded or collapsed
+   */
+  toggleRowExpansion,
+  /**
+   * @description clear sorting, restore data to the original order
+   */
+  clearSort,
+  /**
+   * @description refresh the layout of Table. When the visibility of Table changes, you may need to call this method to get a correct layout
+   */
+  doLayout,
+  /**
+   * @description sort Table manually. Property `prop` is used to set sort column, property `order` is used to set sort order
+   */
+  sort,
+  /**
+   * @description used in lazy Table, must set `rowKey`, update key children
+   */
+  updateKeyChildren,
+  t,
+  setDragVisible,
+  context,
+  computedSumText,
+  computedEmptyText,
+  computedTooltipEffect,
+  computedTooltipOptions,
+  tableLayout,
+  scrollbarViewStyle,
+  scrollbarStyle,
+  scrollBarRef,
+  /**
+   * @description scrolls to a particular set of coordinates
+   */
+  scrollTo,
+  /**
+   * @description set horizontal scroll position
+   */
+  setScrollLeft,
+  /**
+   * @description set vertical scroll position
+   */
+  setScrollTop,
+  /**
+   * @description whether to allow drag the last column
+   */
+  allowDragLastColumn: props.allowDragLastColumn,
 })
 </script>
