+++ packages/components/table/src/table/defaults.ts
@@ -104,38 +104,128 @@ type CellStyle<T extends DefaultRow> =
       columnIndex: number
     }) => CSSProperties)
 type Layout = 'fixed' | 'auto'
-interface TableProps<T extends DefaultRow> {
-  data: T[]
+interface TableProps<T extends DefaultRow = DefaultRow> {
+  /**
+   * @description table data
+   */
+  data?: T[]
+  /**
+   * @description size of Table
+   */
   size?: ComponentSize
   width?: string | number
+  /**
+   * @description table's height. By default it has an `auto` height. If its value is a number, the height is measured in pixels; if its value is a string, the value will be assigned to element's style.height, the height is affected by external styles
+   */
   height?: string | number
+  /**
+   * @description table's max-height. The legal value is a number or the height in px
+   */
   maxHeight?: string | number
+  /**
+   * @description whether width of column automatically fits its container
+   */
   fit?: boolean
+  /**
+   * @description whether Table is striped
+   */
   stripe?: boolean
+  /**
+   * @description whether Table has vertical border
+   */
   border?: boolean
+  /**
+   * @description key of row data, used for optimizing rendering. Required if `reserve-selection` is on or display tree data. When its type is String, multi-level access is supported, e.g. `user.info.id`, but `user.info[0].id` is not supported, in which case `Function` should be used
+   */
   rowKey?: string | ((row: T) => string)
   context?: Table<T>
+  /**
+   * @description whether Table header is visible
+   */
   showHeader?: boolean
+  /**
+   * @description whether to display a summary row
+   */
   showSummary?: boolean
+  /**
+   * @description displayed text for the first column of summary row
+   */
   sumText?: string
+  /**
+   * @description custom summary method
+   */
   summaryMethod?: SummaryMethod<T>
+  /**
+   * @description function that returns custom class names for a row, or a string assigning class names for every row
+   */
   rowClassName?: ColumnCls<T>
+  /**
+   * @description function that returns custom style for a row, or an object assigning custom style for every row
+   */
   rowStyle?: ColumnStyle<T>
+  /**
+   * @description function that returns custom class names for a cell, or a string assigning class names for every cell
+   */
   cellClassName?: CellCls<T>
+  /**
+   * @description function that returns custom style for a cell, or an object assigning custom style for every cell
+   */
   cellStyle?: CellStyle<T>
+  /**
+   * @description function that returns custom class names for a row in table header, or a string assigning class names for every row in table header
+   */
   headerRowClassName?: ColumnCls<T>
+  /**
+   * @description function that returns custom style for a row in table header, or an object assigning custom style for every row in table header
+   */
   headerRowStyle?: ColumnStyle<T>
+  /**
+   * @description function that returns custom class names for a cell in table header, or a string assigning class names for every cell in table header
+   */
   headerCellClassName?: CellCls<T>
+  /**
+   * @description function that returns custom style for a cell in table header, or an object assigning custom style for every cell in table header
+   */
   headerCellStyle?: CellStyle<T>
+  /**
+   * @description whether current row is highlighted
+   */
   highlightCurrentRow?: boolean
+  /**
+   * @description key of current row, a set only prop
+   */
   currentRowKey?: string | number
+  /**
+   * @description displayed text when data is empty. You can customize this area with `#empty`
+   */
   emptyText?: string
+  /**
+   * @description set expanded rows by this prop, prop's value is the keys of expand rows, you should set row-key before using this prop
+   */
   expandRowKeys?: Array<string>
+  /**
+   * @description whether expand all rows by default, works when the table has a column type="expand" or contains tree structure data
+   */
   defaultExpandAll?: boolean
+  /**
+   * @description enable expandable rows, works when the table has a column type="expand"
+   */
   rowExpandable?: (row: T, index: number) => boolean
+  /**
+   * @description set the default sort column and order. property `prop` is used to set default sort column, property `order` is used to set default sort order
+   */
   defaultSort?: Sort
+  /**
+   * @description the `effect` of the overflow tooltip
+   */
   tooltipEffect?: string
+  /**
+   * @description the options for the overflow tooltip, [see the following tooltip component](tooltip.html#attributes)
+   */
   tooltipOptions?: TableOverflowTooltipOptions
+  /**
+   * @description method that returns rowspan and colspan
+   */
   spanMethod?: (data: {
     row: T
     rowIndex: number
@@ -148,20 +238,61 @@ interface TableProps<T extends DefaultRow> {
         colspan: number
       }
     | undefined
+  /**
+   * @description controls the behavior of master checkbox in multi-select tables when only some rows are selected (but not all). If true, all rows will be selected, else deselected
+   */
   selectOnIndeterminate?: boolean
+  /**
+   * @description horizontal indentation of tree data
+   */
   indent?: number
+  /**
+   * @description configuration for rendering nested data
+   */
   treeProps?: TreeProps
+  /**
+   * @description whether to lazy loading data
+   */
   lazy?: boolean
+  /**
+   * @description method for loading child row data, only works when `lazy` is true
+   */
   load?: (row: T, treeNode: TreeNode, resolve: (data: T[]) => void) => void
   className?: string
   style?: StyleValue
+  /**
+   * @description sets the algorithm used to lay out table cells, rows, and columns
+   */
   tableLayout?: Layout
+  /**
+   * @description always show scrollbar
+   */
   scrollbarAlwaysOn?: boolean
+  /**
+   * @description ensure main axis minimum-size doesn't follow the content
+   */
   flexible?: boolean
+  /**
+   * @description whether to hide extra content and show them in a tooltip when hovering on the cell.It will affect all the table columns
+   */
   showOverflowTooltip?: boolean | TableOverflowTooltipOptions
+  /**
+   * @description function that formats cell tooltip content, works when `show-overflow-tooltip` is `true`
+   */
   tooltipFormatter?: TableOverflowTooltipFormatter<T>
   appendFilterPanelTo?: string
   scrollbarTabindex?: number | string
+  /**
+   * @description whether to allow drag the last column
+   */
+  allowDragLastColumn?: boolean
+  /**
+   * @description whether to preserve expanded row content in DOM when collapsed
+   */
+  preserveExpandedContent?: boolean
+  /**
+   * @description whether to use native scrollbars
+   */
   nativeScrollbar?: boolean
 }
 
@@ -210,6 +341,9 @@ interface TableConfigContext {
   tooltipFormatter?: TableOverflowTooltipFormatter<any>
 }
 
+/**
+ * @deprecated Removed after 3.0.0, Use `TableProps` instead.
+ */
 export default {
   /**
    * @description table data
