+++ packages/components/table/src/util.ts
@@ -21,6 +21,15 @@ import type { DefaultRow, Table, TreeProps } from './table/defaults'
 import type { TableColumnCtx } from './table-column/defaults'
 import type { CSSProperties, VNode, VNodeArrayChildren } from 'vue'
 
+let tableIdSeed = 1
+let columnIdSeed = 1
+
+export const createTableId = (namespace: string) =>
+  `${namespace}-table_${tableIdSeed++}`
+
+export const createTableColumnId = (parentId: string) =>
+  `${parentId}_column_${columnIdSeed++}`
+
 export type TableOverflowTooltipOptions = Partial<
   Pick<
     ElTooltipProps,
